Post zip url and access key to the parent window
This change will post zip url and access key to the parent window in case CP is iframed. Please let me know if I should be more aggressive specifying the targets, currently I will need it to work with my canary instance and localhost, not sure if it is safe to just hard code them though. Thanks

Reviewed in https://codereview.appspot.com/93630043/